clears framebuffer
NOTES
Adds useful reference.

Shows useful info about the framebuffer device on the console.

Fills the entire framebuffer with zeros, to achieve this we are
mapping the framebuffer memory via mmap() as mentioned in the
Linux Kernel documentation (that applications probably want to
use mmap to operate on the framebuffer).

Note that for this to work you must be in console mode, not the
usual desktop environment or X, via Ctrl+Alt+FKEY. Also note that
the code, at this point, is not negotiating with the console so
both are writing to the framebuffer and thus the screen will
flicker.