draws the player
NOTES
We get our player centered at the screen

We get a flickering screen but that was expected because we are
not using a buffering strategy, we are writing to the framebuffer
directly. What happens is that we have no control over the hardware
and so it decides when to update the screen. This could in part
explain the flickering but it could be the combined effect of
having a console that also writes to the framebuffer.

There are no padding bytes in my screen, this is why we can omit
the padding in the computation of the offset `of` when drawing
the player.